Fix for enum type duplication in graphene schema
The converter code was created a new type for each Enum column that it
encountered. The problem is where the same Enum type is used for
multiple columns. In that case graphene will notice two schema types
with the same underlying name and throw an Assertion error like:

AssertionError: Found different types with the same name in the schema: pet_kind, pet_kind.

This fix creates a new registry dict for storing a type.name to Enum
mapping that allows for reuse.