Removed assertions that were hiding exceptions
The two assertions that were testing the sqlalchemy classes and instances
worked by catching and silencing exceptions. This could cause debugging
problems like:

https://github.com/graphql-python/graphene-sqlalchemy/issues/121
https://github.com/graphql-python/graphene-sqlalchemy/issues/122

on the other hand without them, if someone uses unmapped class/model it
would be trivial to diagnose the problem. So I believe we would be
better off without them.